xray-analyzer/backend/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import io
import torch
import torchvision.transforms as transforms
from PIL import Image
import json
import logging

# Ensure logging is setup
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(title="MedGemma Omni-XRay Inference API", version="1.0.0")

# CORS for frontend React Application to hit this backend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"], 
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Simulated initialization of the MedGemma + Vision Encoder
# In production, models like `torchxrayvision` or `transformers` 
# would be loaded into VRAM here natively.
model_loaded = False
vision_encoder = None
medgemma_llm = None

@app.on_event("startup")
async def load_models():
    global model_loaded
    # logger.info("Loading Vision Encoder (e.g., DenseNet121 from torchxrayvision)...")
    # vision_encoder = xrv.models.DenseNet(weights="densenet121-res224-all")
    # logger.info("Loading MedGemma-1.5-4b-it via PEFT/transformers...")
    # medgemma_llm = AutoModelForCausalLM.from_pretrained(...)
    logger.info("Models pseudo-loaded into memory (Simulated for this script).")
    model_loaded = True

@app.post("/api/diagnose")
async def analyze_xray(file: UploadFile = File(...), scan_type: str = 'chest'):
    if not model_loaded:
        raise HTTPException(status_code=503, detail="Models are still loading into VRAM.")

    contents = await file.read()
    try:
        image = Image.open(io.BytesIO(contents)).convert("RGB")
    except Exception as e:
        raise HTTPException(status_code=400, detail="Invalid Image Format. Must be decodable picture file.")

    logger.info(f"Processing a {scan_type.upper()} radiograph of size {image.size}...")

    # Real Inference Pipeline (Pseudocode representation of logic):
    # 1. Image Preprocessing
    # transform = transforms.Compose([transforms.Resize(224), transforms.ToTensor()])
    # tensor = transform(image).unsqueeze(0)
    
    # 2. Vision Embeddings
    # with torch.no_grad():
    #     viz_features = vision_encoder.features(tensor)
    
    # 3. Prompting MedGemma
    # prompt = f"Analyze this {scan_type} radiograph and provide findings:"
    # text_output = medgemma_llm.generate(prompt_embeds=viz_features, ...)
    
    # Returning structured AI JSON payload to the frontend
    if scan_type == 'chest':
        return {
            "overall_status": "Abnormal",
            "findings": ["Infiltration", "Pleural Effusion"],
            "description": "Bilateral lung fields demonstrate patchy opacities in the mid-to-lower zones suggestive of infiltration. Blunting of the left costophrenic angle indicates mild pleural effusion. The cardiomediastinal silhouette is within normal limits. Trachea is midline.",
            "confidence": 91.5,
            "bbox": [{"top": 60, "left": 70, "width": 15, "height": 10}, {"top": 40, "left": 30, "width": 20, "height": 20}] # Highlighting lungs
        }
    else:
        return {
            "overall_status": "Abnormal",
            "findings": ["Fracture", "Cortical Disruption"],
            "description": "Evidence of an acute transverse fracture through the diaphysis with 2mm dorsal displacement. No intra-articular extension is apparent. Surrounding soft tissue swelling and joint effusion are evident.",
            "confidence": 96.2,
            "bbox": [{"top": 45, "left": 45, "width": 10, "height": 10}] # Highlighting midshaft bone
        }

@app.get("/health")
def health_check():
    return {"status": "ok", "gpu_available": torch.cuda.is_available()}
