xray-analyzer/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
torch>=2.0.0
torchvision
Pillow
python-multipart
